requirements.txt
scipy==1.6.1
numpy==1.19.2
tqdm==4.56.0
numba==0.52.0
matplotlib==3.3.4
pandas==1.2.3
sktime==0.5.3
tensorflow-gpu==2.3.0
scikit_learn==0.24.1
